tests/unit/requirements.txt
kubernetes>=24.2.0
PyYAML>=3.11
coverage==7.10.7
pytest
pytest-xdist
pytest-forked
ansible-lint
argcomplete
